api/requirements.txt
# Web Framework
litestar[standard]==2.12.1
uvicorn[standard]==0.30.6

# LangChain
langchain==0.3.7
langchain-community==0.3.7

# LLM
ollama==0.4.4

# Embeddings
sentence-transformers==3.3.1

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6

# Document Processing - Básico
PyPDF2==3.0.1
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5

# OCR - Imagens
pytesseract==0.3.10
Pillow==10.1.0

# Word/PowerPoint - Conversão
python-docx==1.1.0
python-pptx==0.6.23
pypandoc==1.13

# Web Scraping
beautifulsoup4==4.12.3
httpx==0.27.2
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1